* Refactor out native entrycomponents * Add flag to HeaderHoc to disable bottom border * Change RPC call to request access / accept invite token
