+++ shared/teams/join-team/index.desktop.js
@@ -1,0 +1,62 @@
+// @flow
+import React from 'react'
+import {Box, Button, Input, Text, ScrollView} from '../../common-adapters/index'
+import {SuccessComponent, styleContainer, stylePadding} from './index.shared'
+import {PopupDialog} from '../../common-adapters/'
+import {globalColors, globalMargins, globalStyles} from '../../styles'
+
+import type {Props} from '.'
+const EntryComponent = ({errorText, name, onNameChange, onSubmit}) => (
+  <Box style={globalStyles.flexBoxColumn}>
+    {!!errorText &&
+      <Box
+        style={{
+          ...styleContainer,
+          backgroundColor: globalColors.red,
+        }}
+      >
+        <Text
+          style={{margin: globalMargins.tiny, textAlign: 'center', width: '100%'}}
+          type="BodySemibold"
+          backgroundMode="HighRisk"
+        >
+          {errorText}
+        </Text>
+      </Box>}
+    <Box
+      style={{
+        ...globalStyles.flexBoxColumn,
+        ...stylePadding,
+        alignItems: 'center',
+        justifyContent: 'center',
+      }}
+    >
+      <Text type="Header">Join a team</Text>
+      <Input
+        autoFocus={true}
+        hintText="Token or team name"
+        value={name}
+        onChangeText={onNameChange}
+        onEnterKeyDown={onSubmit}
+      />
+      <Box style={{...globalStyles.flexBoxRow, marginTop: globalMargins.large}}>
+        <Button type="Primary" style={{marginLeft: globalMargins.tiny}} onClick={onSubmit} label="Continue" />
+      </Box>
+    </Box>
+  </Box>
+)
+
+export const Contents = (props: Props) => (
+  <ScrollView>
+    {!!props.success && SuccessComponent(props)}
+    {!props.success && EntryComponent(props)}
+  </ScrollView>
+)
+
+const PopupWrapped = (props: Props) => (
+  <PopupDialog onClose={props.onBack}>
+    <Contents {...props} />
+  </PopupDialog>
+)
+
+export default PopupWrapped
